+++ examples/starter/src/components/Elements.token.ts
@@ -16,11 +16,9 @@ import { flow } from 'lodash';
 import { addClasses } from '@bodiless/fclasses';
 import {
   asBodilessLink,
-  Editable,
-  withPlaceholder,
+  asEditable as asEditableCore,
 } from '@bodiless/components';
 import { asBodilessImage } from '@bodiless/components-ui';
-import { withChild, withNodeKey } from '@bodiless/core';
 import {
   asAlignCenter,
   asAlignRight,
@@ -61,12 +59,7 @@ const asEditableImage = asBodilessImage;
 const asEditableLink = asBodilessLink;
 
 /* Edit component */
-const asEditable = (nodeKey?: string, placeholder?: string) => withChild(
-  flow(
-    withNodeKey(nodeKey),
-    withPlaceholder(placeholder),
-  )(Editable),
-);
+const asEditable = asEditableCore;
 
 export {
   asBold,