+++ packages/bodiless-components/src/Editable.tsx
@@ -12,22 +12,25 @@
  * limitations under the License.
  */
 
-import React, { ComponentType as CT, ClipboardEvent } from 'react';
-import { flow } from 'lodash';
+import React, { ComponentType as CT, ClipboardEvent, ComponentType } from 'react';
 import ContentEditable from 'react-contenteditable';
 import { observer } from 'mobx-react-lite';
+import { flowRight, pickBy } from 'lodash';
 import {
   withNode,
   useNode,
-  withNodeKey,
   useEditContext,
-  withChild,
+  WithNodeProps,
+  WithNodeKeyProps,
+  withNodeKey,
 } from '@bodiless/core';
 import './Editable.css';
 
 type Props = {
   placeholder?: string;
-};
+  children?: string;
+} & Partial<WithNodeProps>;
+
 type Data = {
   text: string;
 };
@@ -36,7 +39,7 @@ type Data = {
 const Text = observer((props: Props) => {
   const { placeholder } = props;
   const { node } = useNode<Data>();
-  const text = node.data.text || placeholder || '';
+  const text = (node.data.text !== undefined ? node.data.text : props.children) || placeholder || '';
   // eslint-disable-next-line react/no-danger
   return <span dangerouslySetInnerHTML={{ __html: text }} />;
 });
@@ -55,7 +58,7 @@ const EditableText = observer((props: Props) => {
   };
   const { placeholder } = props;
   const placeholderDataAttr = placeholder || '';
-  const text = node.data.text || '';
+  const text = (node.data.text !== undefined ? node.data.text : props.children) || '';
   return (
     <ContentEditable
       tagName="span"
@@ -84,14 +87,48 @@ const withPlaceholder = <P extends object> (placeholder?: string) => (Component:
   return WithPlaceholder;
 };
 
-const asEditable = (nodeKey?: string, placeholder?: string) => (
-  withChild(
-    flow(
-      withNodeKey(nodeKey),
+/**
+ * asEditable takes a nodeKey and a placeholder, and returns an HOC which injects
+ * an editable span as a child of the wrapped component.  The original children
+ * of the wrapped component will become children of the editable span.  In addition,
+ * `nodeKey`, `nodeCollection` and `placeholder` props passed to the enhanced
+ * component will be forwarded to the editable span.
+ *
+ * @param nodeKeys A nodeKey or an object containing nodeKey and nodeCollection.
+ * @param placeholder A string to use as placeholder text.
+ * @return A HOC to inject an editable span.
+ */
+const asEditable = (nodeKeys?: WithNodeKeyProps, placeholder?: string) => (
+  <P extends object>(Component: CT<P>|string) => {
+    const EditableChild = flowRight(
+      withNodeKey(nodeKeys),
       withPlaceholder(placeholder),
-    )(Editable),
-  )
+    )(Editable);
+    const AsEditable = (props: P & Props) => {
+      // @TODO: Improve `withChild` to allow this kind of prop splitting.
+      const {
+        children,
+        nodeKey,
+        nodeCollection,
+        placeholder: placeholderProp,
+        ...rest
+      } = props;
+      const editableProps = pickBy({
+        children,
+        nodeKey,
+        nodeCollection,
+        placeholder: placeholderProp,
+      });
+      return (
+        <Component {...rest as P}>
+          <EditableChild {...editableProps} />
+        </Component>
+      );
+    };
+    return AsEditable as ComponentType<P & Props>;
+  }
 );
+
 export default Editable;
 export {
   withPlaceholder,