+++ packages/bodiless-core-ui/src/LocalContextMenu.tsx
@@ -22,7 +22,7 @@ import {
 import {
   ComponentFormTitle, ComponentFormLabel, ComponentFormText, ComponentFormButton,
   ComponentFormCloseButton, ComponentFormSubmitButton, Icon, Div, ToolbarButton,
-  ComponentFormUnwrapButton, ComponentFormDescription,
+  ComponentFormUnwrapButton, ComponentFormTextArea, ComponentFormDescription,
 } from '@bodiless/ui';
 import ReactTagsField from './ReactTags';
 
@@ -41,6 +41,7 @@ const LocalTooltip: FC<ReactTooltip['props']> = props => (
 
 const ui = {
   ComponentFormText,
+  ComponentFormTextArea,
   ComponentFormButton,
   ComponentFormCloseButton,
   ComponentFormUnwrapButton,