+++ packages/bodiless-core/src/contextMenuForm.tsx
@@ -12,9 +12,9 @@
  * limitations under the License.
  */
 
-import React, { ReactNode } from 'react';
+import React, { ReactNode, useCallback } from 'react';
 import {
-  Form, FormApi, FormState, Text,
+  Form, FormApi, FormState, Text, TextArea,
 } from 'informed';
 import { UI } from './Types/ContextMenuTypes';
 import ReactTagsField from './components/ReactTagsField';
@@ -28,6 +28,7 @@ const defaultUI = {
   ComponentFormSubmitButton: 'button',
   ComponentFormUnwrapButton: 'button',
   ComponentFormText: Text,
+  ComponentFormTextArea: TextArea,
   ComponentFormError: 'div',
   Form: 'div',
   ReactTags: ReactTagsField,
@@ -38,67 +39,98 @@ const defaultUI = {
 
 export const getUI = (ui: UI = {}) => ({ ...defaultUI, ...ui });
 
+export type Options<D> = {
+  submitValues?: (componentData: D) => boolean|void;
+  onClose?: (componentData: D) => boolean|void;
+  initialValues?: D;
+  hasSubmit?: Boolean;
+};
+
 export type FormProps = {
   closeForm: () => void;
   ui?: UI;
   'aria-label'?: string;
 };
 
-export type FormBodyProps<D> = FormProps & {
+export type FormBodyProps<D> = FormProps & Options<D> & {
   formApi: FormApi<D>;
   formState: FormState<D>;
 };
 
 export type FormBodyRenderer<D> = (props: FormBodyProps<D>) => ReactNode;
-
-export type Options<D> = {
-  submitValues?: (componentData: D) => boolean|void;
-  initialValues?: D;
-  hasSubmit?: Boolean;
+type Props<D> = FormProps & Options<D> & {
+  children: FormBodyRenderer<D>|ReactNode,
 };
 
-const contextMenuForm = <D extends object>(options: Options<D>) => (
-  renderFormBody: FormBodyRenderer<D>,
-) => {
-  const ContextMenuForm = ({ closeForm, ui, ...rest }: FormProps) => {
-    const { ComponentFormCloseButton, ComponentFormSubmitButton } = getUI(ui);
-    const { submitValues, initialValues, hasSubmit = true } = options;
-    return (
-      <Form
-        onSubmit={(values: D) => {
-          if (!submitValues || !submitValues(values)) {
-            closeForm();
-          }
-        }}
-        initialValues={initialValues}
-        {...rest}
-      >
-        {({ formApi, formState }) => (
-          <>
-            <ComponentFormCloseButton
-              type="button"
-              onClick={closeForm}
-              aria-label="Cancel"
-            />
-            {renderFormBody({
-              closeForm,
-              formApi,
-              formState,
-              ui,
-            })}
-            {
-              hasSubmit && !formState.invalid
-              && (
-                <ComponentFormSubmitButton aria-label="Submit" />
-              )
-            }
-          </>
-        )}
-      </Form>
-    );
+export const ContextMenuForm = <D extends object>({
+  closeForm,
+  onClose,
+  ui,
+  submitValues = () => undefined,
+  initialValues = {} as D,
+  hasSubmit = true,
+  children = () => <></>,
+  ...rest
+}: Props<D>) => {
+  const { ComponentFormCloseButton, ComponentFormSubmitButton } = getUI(ui);
+  const callOnClose = (values: D) => {
+    if (typeof onClose === 'function') {
+      onClose(values);
+    }
+    closeForm();
   };
-  ContextMenuForm.displayName = 'ComponentForm';
-  return ContextMenuForm;
+  return (
+    <Form
+      onSubmit={(values: D) => {
+        if (!submitValues(values)) {
+          callOnClose(values);
+        }
+      }}
+      initialValues={initialValues}
+      {...rest}
+    >
+      {({ formApi, formState }) => (
+        <>
+          <ComponentFormCloseButton
+            type="button"
+            onClick={() => callOnClose(formState.values)}
+            aria-label="Cancel"
+          />
+          {typeof children === 'function'
+            ? children({
+              closeForm, formApi, formState, ui,
+            })
+            : children}
+          {hasSubmit && !formState.invalid
+          && (
+            <ComponentFormSubmitButton aria-label="Submit" />
+          )}
+        </>
+      )}
+    </Form>
+  );
 };
 
-export default contextMenuForm;
+export const contextMenuForm = <D extends object>(options: Options<D> = {}) => (
+  renderFormBody?: FormBodyRenderer<D>,
+) => (
+  (props: Omit<Props<D>, 'children'>) => (
+    <ContextMenuForm {...options} {...props}>
+      {renderFormBody || (() => <></>)}
+    </ContextMenuForm>
+  )
+);
+
+type HookOptions<D> = Options<D> & {
+  renderFormBody?: FormBodyRenderer<D>,
+};
+
+const useContextMenuForm = <D extends object>(options: HookOptions<D> = {}) => {
+  const { renderFormBody, ...rest } = options;
+  return useCallback(
+    contextMenuForm(rest)(renderFormBody),
+    [options],
+  );
+};
+
+export default useContextMenuForm;