+++ packages/bodiless-layouts/src/ComponentSelector/uiContext.tsx
@@ -12,7 +12,6 @@
  * limitations under the License.
  */
 
-
 import React from 'react';
 import { FinalUI } from './types';